+++ Pangle/source/plugin/Assets/GoogleMobileAds/Mediation/Pangle/Platforms/iOS/Externs.cs
@@ -22,9 +22,6 @@ namespace GoogleMobileAds.Mediation.Pangle.iOS
     internal class Externs
     {
         [DllImport("__Internal")]
-        internal static extern void GADUMPangleSetGDPRConsent(int gdpr);
-
-        [DllImport("__Internal")]
         internal static extern void GADUMPangleSetPAConsent(int paConsent);
     }
 }