+++ Pangle/source/plugin/Assets/GoogleMobileAds/Mediation/Pangle/Platforms/iOS/PangleClient.cs
@@ -14,6 +14,8 @@
 
 #if UNITY_IOS
 
+using UnityEngine;
+
 using GoogleMobileAds.Mediation.Pangle.Common;
 
 namespace GoogleMobileAds.Mediation.Pangle.iOS
@@ -33,7 +35,9 @@ namespace GoogleMobileAds.Mediation.Pangle.iOS
 
         public void SetGDPRConsent(int gdpr)
         {
-            Externs.GADUMPangleSetGDPRConsent(gdpr);
+            MonoBehaviour.print("[Pangle Plugin] 'SetGDPRConsent' is now deprecated and will be " +
+                    "removed in a future release. GDPR consent is now automatically read by the " +
+                    "Pangle SDK.");
         }
 
         public void SetPAConsent(int paConsent)